Only use dynamic paths in virtualenv when creating the pyenv package
Summary:
This is a dev productivity change only and will not impact customer packages.

When running `install_python_requirements.sh`, setting the absolute paths inside the virtual env
to be dynamic paths requires searching through each file, which can take a while. This diff
makes it so this is only done when `--create_package` is specified, at which point it will remove
the old pyenv and create a new one.

Test Plan: bin/install_python_requirements.sh

Reviewers: daniel, arnav, sanketh

Reviewed By: sanketh

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D9914

+++ managed/devops/bin/install_python_requirements.sh
@@ -46,6 +46,11 @@ while [[ $# -gt 0 ]]; do
   shift
 done
 
+if [[ "$should_create_package" ]]; then
+  log "Removing any existing pyenv"
+  rm -rf "$virtualenv_dir"
+fi
+
 if should_use_virtual_env; then
   activate_virtualenv
   fix_virtualenv_permissions
@@ -65,18 +70,20 @@ else
   log "Installing ybops package"
   install_ybops_package
 
-  log "Changing virtualenv absolute paths to dynamic paths"
-  # Change shebangs to use local python instead of absolute python path - required for our Jenkins
-  # pipeline and packaging (e.g. "/tmp/python_virtual_env/bin/python" -> "/usr/bin/env python")
-  LC_ALL=C find $virtualenv_dir ! -name '*.pyc' -type f -exec sed -i.yb_tmp \
-    -e "1s|${virtualenv_dir}/bin/python|/usr/bin/env python|" {} \; -exec rm {}.yb_tmp \;
+  if [[ "$should_create_package" ]]; then
+    log "Changing virtualenv absolute paths to dynamic paths"
+    # Change shebangs to use local python instead of absolute python path - required for our Jenkins
+    # pipeline and packaging (e.g. "/tmp/python_virtual_env/bin/python" -> "/usr/bin/env python")
+    LC_ALL=C find $virtualenv_dir ! -name '*.pyc' -type f -exec sed -i.yb_tmp \
+      -e "1s|${virtualenv_dir}/bin/python|/usr/bin/env python|" {} \; -exec rm {}.yb_tmp \;
 
-  # Change VIRTUAL_ENV variable to be dynamic. Instead of hardcoding the path to the vitualenv
-  # directory, the VIRTUAL_ENV variable should print the filepath of the directory two above it.
-  new_venv_assignment='VIRTUAL_ENV="$(cd "$(dirname "$(dirname "${BASH_SOURCE[0]}" )")" \&\& pwd -P)"'
-  sed -i.yb_tmp \
-    -e "s|VIRTUAL_ENV=\"${virtualenv_dir}\"|${new_venv_assignment}|" $virtualenv_dir/bin/activate
-  rm $virtualenv_dir/bin/activate.yb_tmp
+    # Change VIRTUAL_ENV variable to be dynamic. Instead of hardcoding the path to the vitualenv
+    # directory, the VIRTUAL_ENV variable should print the filepath of the directory two above it.
+    new_venv_assignment='VIRTUAL_ENV="$(cd "$(dirname "$(dirname "${BASH_SOURCE[0]}" )")" \&\& pwd -P)"'
+    sed -i.yb_tmp \
+      -e "s|VIRTUAL_ENV=\"${virtualenv_dir}\"|${new_venv_assignment}|" $virtualenv_dir/bin/activate
+    rm $virtualenv_dir/bin/activate.yb_tmp
+  fi
 
   if should_use_virtual_env; then
     log "Expecting there to be no differences between the output of 'pip freeze' and the contents" \
